requirements.txt
langchain-google-genai>=2.1.0
langchain-community
faiss-cpu
sentence-transformers
PyPDF2=3.0.1